Conflicts:
	cli/lib/kontena/cli/apps/logs_command.rb

+++ server/app/jobs/node_cleanup_job.rb
@@ -4,7 +4,7 @@ require_relative '../services/logging'
 class NodeCleanupJob
   include Celluloid
   include Logging
-  include DistributedLocks
+  include CurrentLeader
 
   def initialize
     async.perform
@@ -13,29 +13,25 @@ class NodeCleanupJob
   def perform
     info 'starting to cleanup stale connections'
     every(1.minute.to_i) do
-      cleanup_stale_connections
+      cleanup_stale_connections if leader?
     end
 
     info 'starting to cleanup stale nodes'
     every(1.hour.to_i) do
-      cleanup_stale_nodes
+      cleanup_stale_nodes if leader?
     end
   end
 
   def cleanup_stale_connections
-    with_dlock('node_cleanup_job:stale_connections', 0) do
-      HostNode.where(:last_seen_at.lt => 1.minute.ago).each do |node|
-        node.set(connected: false)
-      end
+    HostNode.where(:last_seen_at.lt => 1.minute.ago).each do |node|
+      node.set(connected: false)
     end
   end
 
   def cleanup_stale_nodes
-    with_dlock('node_cleanup_job:stale_nodes', 0) do
-      HostNode.where(:last_seen_at.lt => 1.hour.ago).each do |node|
-        unless node.connected?
-          node.destroy
-        end
+    HostNode.where(:last_seen_at.lt => 1.hour.ago).each do |node|
+      if !node.grid.initial_node?(node) && !node.connected? && !node.stateful?
+        node.destroy
       end
     end
   end
